requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
tensorflow==2.18.1
opencv-python==4.8.0.74
numpy==1.24.3
python-dotenv>=1.0.0
pymongo==4.4.1
pydicom==2.4.3
pillow>=10.0.0
pytest==7.4.0
python-jose[cryptography]==3.3.0
passlib==1.7.4